sfd
fsd

+++ routes/user.js
@@ -345,7 +345,9 @@ router.put("/oneuser/:id",ensureToken, (req, res) => {
  pool.query("SELECT * FROM users", (err, result) => {
         if (!err) {
             var a=result.rows.filter(item=>item.id==req.params.id) 
-            fs.unlink(`./Images/${a[0].image}`,()=>{})
+            if(a[0].image){
+                fs.unlink(`./Images/${a[0].image}`,()=>{})
+            }
     pool.query(
     'UPDATE users SET address = $1,description=$2,email=$3, image=$4,last_name=$5,phone_number=$6,username=$7 WHERE id = $8',
         [body.address, body.description, body.email,imgName,body.last_name,body.phone_number,body.username,id],
@@ -363,7 +365,8 @@ router.put("/oneuser/:id",ensureToken, (req, res) => {
       pool.query("SELECT * FROM users", (err, result) => {
              if (!err) {
                  var a=result.rows.filter(item=>item.id==req.params.id) 
-                 fs.unlink(`./Images/${a[0].image}`,()=>{})
+               if(a[0].image){ fs.unlink(`./Images/${a[0].image}`,()=>{})}
+                
          pool.query(
          'UPDATE users SET address = $1,description=$2,email=$3, image=$4,last_name=$5,phone_number=$6,username=$7 WHERE id = $8',
              [body.address, body.description, body.email,body.image,body.last_name,body.phone_number,body.username,id],
